hardware/rpi_server/requirements.txt
# 4DX@HOME Raspberry Pi Server - Dependencies
# Python 3.9+

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# WebSocket Client
websockets==12.0

# MQTT Client
paho-mqtt==1.6.1

# Environment & Configuration
python-dotenv==1.0.0

# Async I/O
aiofiles==23.2.1

# Logging & Utilities
coloredlogs==15.0.1
python-json-logger==2.0.7

# Date & Time
python-dateutil==2.8.2

# Testing (optional, for development)
pytest==7.4.3
pytest-asyncio==0.21.1
